+++ test/AbstractPluginManagerTest.php
@@ -1,21 +1,20 @@
 <?php
+
 /**
- * Zend Framework (http://framework.zend.com/)
- *
- * @link      http://github.com/zendframework/zf2 for the canonical source repository
- * @copyright Copyright (c) 2005-2015 Zend Technologies USA Inc. (http://www.zend.com)
- * @license   http://framework.zend.com/license/new-bsd New BSD License
+ * @see       https://github.com/laminas/laminas-servicemanager for the canonical source repository
+ * @copyright https://github.com/laminas/laminas-servicemanager/blob/master/COPYRIGHT.md
+ * @license   https://github.com/laminas/laminas-servicemanager/blob/master/LICENSE.md New BSD License
  */
 
-namespace ZendTest\ServiceManager;
+namespace LaminasTest\ServiceManager;
 
+use Laminas\ServiceManager\Config;
+use Laminas\ServiceManager\Exception\RuntimeException;
+use Laminas\ServiceManager\ServiceManager;
+use LaminasTest\ServiceManager\TestAsset\FooPluginManager;
+use LaminasTest\ServiceManager\TestAsset\MockSelfReturningDelegatorFactory;
 use ReflectionClass;
 use ReflectionObject;
-use Zend\ServiceManager\Config;
-use Zend\ServiceManager\Exception\RuntimeException;
-use Zend\ServiceManager\ServiceManager;
-use ZendTest\ServiceManager\TestAsset\FooPluginManager;
-use ZendTest\ServiceManager\TestAsset\MockSelfReturningDelegatorFactory;
 
 class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
 {
@@ -34,7 +33,7 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
         $this->serviceManager = new ServiceManager();
         $this->pluginManager = new FooPluginManager(new Config([
             'factories' => [
-                'Foo' => 'ZendTest\ServiceManager\TestAsset\FooFactory',
+                'Foo' => 'LaminasTest\ServiceManager\TestAsset\FooFactory',
             ],
             'shared' => [
                 'Foo' => false,
@@ -46,7 +45,7 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
     {
         $pluginManager = new FooPluginManager(new Config([
             'factories' => [
-                'Foo' => 'ZendTest\ServiceManager\TestAsset\FooFactory'
+                'Foo' => 'LaminasTest\ServiceManager\TestAsset\FooFactory'
             ],
             'shared' => [
                 'Foo' => false
@@ -63,13 +62,13 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
         $pluginManager->get('Foo', ['key1' => 'value1']);
 
         $value = $reflProperty->getValue($pluginManager);
-        $this->assertInstanceOf('ZendTest\ServiceManager\TestAsset\FooFactory', $value['foo']);
+        $this->assertInstanceOf('LaminasTest\ServiceManager\TestAsset\FooFactory', $value['foo']);
         $this->assertEquals(['key1' => 'value1'], $value['foo']->getCreationOptions());
 
         $pluginManager->get('Foo', ['key2' => 'value2']);
 
         $value = $reflProperty->getValue($pluginManager);
-        $this->assertInstanceOf('ZendTest\ServiceManager\TestAsset\FooFactory', $value['foo']);
+        $this->assertInstanceOf('LaminasTest\ServiceManager\TestAsset\FooFactory', $value['foo']);
         $this->assertEquals(['key2' => 'value2'], $value['foo']->getCreationOptions());
     }
 
@@ -78,7 +77,7 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
      */
     public function testGetFaultyRegisteredInvokableThrowsException()
     {
-        $this->setExpectedException('Zend\ServiceManager\Exception\ServiceNotFoundException');
+        $this->setExpectedException('Laminas\ServiceManager\Exception\ServiceNotFoundException');
 
         $pluginManager = new FooPluginManager();
         $pluginManager->setInvokableClass('helloWorld', 'IDoNotExist');
@@ -88,7 +87,7 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
     public function testAbstractFactoryWithMutableCreationOptions()
     {
         $creationOptions = ['key1' => 'value1'];
-        $mock = 'ZendTest\ServiceManager\TestAsset\AbstractFactoryWithMutableCreationOptions';
+        $mock = 'LaminasTest\ServiceManager\TestAsset\AbstractFactoryWithMutableCreationOptions';
         $abstractFactory = $this->getMock($mock, ['setCreationOptions']);
         $abstractFactory->expects($this->once())
             ->method('setCreationOptions')
@@ -101,7 +100,7 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testMutableMethodNeverCalledWithoutCreationOptions()
     {
-        $mock = 'ZendTest\ServiceManager\TestAsset\CallableWithMutableCreationOptions';
+        $mock = 'LaminasTest\ServiceManager\TestAsset\CallableWithMutableCreationOptions';
         $callable = $this->getMock($mock, ['setCreationOptions']);
         $callable->expects($this->never())
             ->method('setCreationOptions');
@@ -116,7 +115,7 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
     public function testCallableObjectWithMutableCreationOptions()
     {
         $creationOptions = ['key1' => 'value1'];
-        $mock = 'ZendTest\ServiceManager\TestAsset\CallableWithMutableCreationOptions';
+        $mock = 'LaminasTest\ServiceManager\TestAsset\CallableWithMutableCreationOptions';
         $callable = $this->getMock($mock, ['setCreationOptions']);
         $callable->expects($this->once())
             ->method('setCreationOptions')
@@ -135,8 +134,8 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testValidatePluginIsCalledWithDelegatorFactoryIfItsAService()
     {
-        $pluginManager = $this->getMockForAbstractClass('Zend\ServiceManager\AbstractPluginManager');
-        $delegatorFactory = $this->getMock('Zend\\ServiceManager\\DelegatorFactoryInterface');
+        $pluginManager = $this->getMockForAbstractClass('Laminas\ServiceManager\AbstractPluginManager');
+        $delegatorFactory = $this->getMock('Laminas\\ServiceManager\\DelegatorFactoryInterface');
 
         $pluginManager->setService('delegator-factory', $delegatorFactory);
         $pluginManager->addDelegator('foo-service', 'delegator-factory');
@@ -150,9 +149,9 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testSingleDelegatorUsage()
     {
-        $delegatorFactory = $this->getMock('Zend\\ServiceManager\\DelegatorFactoryInterface');
-        /* @var $pluginManager \Zend\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject */
-        $pluginManager = $this->getMockForAbstractClass('Zend\ServiceManager\AbstractPluginManager');
+        $delegatorFactory = $this->getMock('Laminas\\ServiceManager\\DelegatorFactoryInterface');
+        /* @var $pluginManager \Laminas\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject */
+        $pluginManager = $this->getMockForAbstractClass('Laminas\ServiceManager\AbstractPluginManager');
         $realService = $this->getMock('stdClass', [], [], 'RealService');
         $delegator = $this->getMock('stdClass', [], [], 'Delegator');
 
@@ -187,8 +186,8 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
 
     public function testMultipleDelegatorsUsage()
     {
-        /* @var $pluginManager \Zend\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject */
-        $pluginManager = $this->getMockForAbstractClass('Zend\ServiceManager\AbstractPluginManager');
+        /* @var $pluginManager \Laminas\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject */
+        $pluginManager = $this->getMockForAbstractClass('Laminas\ServiceManager\AbstractPluginManager');
 
         $fooDelegator = new MockSelfReturningDelegatorFactory();
         $barDelegator = new MockSelfReturningDelegatorFactory();
@@ -216,15 +215,15 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
         $sm = new ServiceManager();
         $sm->setService('bar', new \stdClass());
 
-        /** @var \Zend\ServiceManager\AbstractPluginManager $pluginManager */
+        /** @var \Laminas\ServiceManager\AbstractPluginManager $pluginManager */
         $pluginManager = new FooPluginManager();
         $pluginManager->setServiceLocator($sm);
 
-        $this->setExpectedException('Zend\ServiceManager\Exception\ServiceLocatorUsageException');
+        $this->setExpectedException('Laminas\ServiceManager\Exception\ServiceLocatorUsageException');
 
         $pluginManager->get('bar');
 
-        $this->fail('A Zend\ServiceManager\Exception\ServiceNotCreatedException is expected');
+        $this->fail('A Laminas\ServiceManager\Exception\ServiceNotCreatedException is expected');
     }
 
     /**
@@ -236,8 +235,8 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
 
         $sm->setInvokableClass('stdClass', 'stdClass');
 
-        /** @var \Zend\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject $pluginManager */
-        $pluginManager = $this->getMockForAbstractClass('Zend\ServiceManager\AbstractPluginManager');
+        /** @var \Laminas\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject $pluginManager */
+        $pluginManager = $this->getMockForAbstractClass('Laminas\ServiceManager\AbstractPluginManager');
 
         $pluginManager
             ->expects($this->once())
@@ -247,7 +246,7 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
 
         $pluginManager->setServiceLocator($sm);
 
-        $this->setExpectedException('Zend\ServiceManager\Exception\ServiceLocatorUsageException');
+        $this->setExpectedException('Laminas\ServiceManager\Exception\ServiceLocatorUsageException');
 
         $pluginManager->get('stdClass');
     }
@@ -257,8 +256,8 @@ class AbstractPluginManagerTest extends \PHPUnit_Framework_TestCase
      */
     public function testWillResetAutoInvokableServiceIfNotValid()
     {
-        /** @var \Zend\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject $pluginManager */
-        $pluginManager = $this->getMockForAbstractClass('Zend\ServiceManager\AbstractPluginManager');
+        /** @var \Laminas\ServiceManager\AbstractPluginManager|\PHPUnit_Framework_MockObject_MockObject $pluginManager */
+        $pluginManager = $this->getMockForAbstractClass('Laminas\ServiceManager\AbstractPluginManager');
 
         $pluginManager
             ->expects($this->any())